+++ examples/print.rs
@@ -3,7 +3,7 @@ use std::io::stdin;
 use termal::reset_terminal;
 
 fn main() {
-    println!("\x1b[52mhello");
-    _ = stdin().read_line(&mut String::new());
+    println!("something hello there\x1b#3");
+    // _ = stdin().read_line(&mut String::new());
     reset_terminal();
 }